+++ XuanYu.Editor.UI/Top/EditToolsModule.axaml.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace XuanYu.Editor.UI;
+
+public partial class EditToolsModule : UserControl
+{
+    public EditToolsModule() => InitializeComponent();
+}