+++ XuanYu.Editor.UI/Top/FileModule.axaml.cs
@@ -1,0 +1,8 @@
+using Avalonia.Controls;
+
+namespace XuanYu.Editor.UI;
+
+public partial class FileModule : UserControl
+{
+    public FileModule() => InitializeComponent();
+}